+++ src/test/java/ConfigurationDecryptTest.java
@@ -11,8 +11,7 @@ public class ConfigurationDecryptTest {
     @Test
     @Ignore
     public void testDectypt() {
-        HttpClient client = new HttpClient();
-        Response response = client.execute("http://host/System/configurationFile?auth=YWRtaW46MTEK");
+        Response response = HttpClient.execute("http://host/System/configurationFile?auth=YWRtaW46MTEK");
         InputStream inputStream = response.body().byteStream();
         System.out.println(ConfigurationDecrypt.decrypt(inputStream));
     }