+++ src/test/java/scanner/brute/basic/BasicAuthScannerTest.java
@@ -1,0 +1,13 @@
+package scanner.brute.basic;
+
+import org.junit.Test;
+
+public class BasicAuthScannerTest {
+
+    @Test
+    public void brute() {
+        BasicAuthScanner ba = new BasicAuthScanner();
+        ba.brute("localhost", new String[] { "12345" });
+    }
+
+}